requirements.txt
scikit-learn==1.4.1
pandas==2.2.1
numpy==1.26.4
mlflow==2.11.1
evidently==0.4.16
azure-ai-ml==1.14.0
azure-identity==1.15.0
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.0
pytest==8.1.0
joblib==1.3.2